meditrack/meditrack/requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
bcrypt==3.2.0
python-dotenv==0.19.0
flask==2.3.2
werkzeug==2.3.7
boto3
